リンクエラーを修正

+++ usr/include/machine/inline/string/strlen.h
@@ -6,7 +6,7 @@
  * Bell V7 string length.
  */
 
-__INLINE int
+__INLINE size_t
 strlen(const char *str) {
 	int rv;
 	const char *strcp = str;
